// frontend/src/app/model/Devis.ts
export class Devis {

    devis_numero: number;
	devis_date: Date;
	devis_numeroproposition: number;
	devis_souscripteur: number;
	devis_objet: string;
	devis_montantnet: number;
	devis_accessoirecompagnie: number;
	devis_accessoireapporteur: number;
	devis_montantfrais: number;
	devis_montanttaxe: number;
	devis_montantttc: number;
	devis_branche: number;
	devis_categorie: number;
	devis_codeutilisateur: number;
	devis_datemodification: Date;
	devis_status: number;
	active: number;


}